+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/S7_ConsoleCommander.lua
@@ -16,12 +16,17 @@ local function S7_FetchPlayers()
     end
 end
 
+--  ################################################################################################################################################
+
+--  ===================
+--  ADD OR REMOVE SKILL
+--  ===================
+
 function S7_AddSkill(...)
     args = {...}
     command = args[1] or ""
     skillName = args[2] or ""
     character = args[3] or ""
-    Ext.Print(skillName .. character)
     if skillName ~= "" and skillName ~= nil then
         S7_FetchPlayers()
         if character == "" or character == nil then
@@ -42,6 +47,34 @@ function S7_AddSkill(...)
     end
 end
 
---  ==================================================
+function S7_RemoveSkill(...)
+    args = {...}
+    command = args[1] or ""
+    skillName = args[2] or ""
+    character = args[3] or ""
+    if skillName ~= "" and skillName ~= nil then
+        S7_FetchPlayers()
+        if character == "" or character == nil then
+            for i, player in ipairs(Osi.DB_IsPlayer:Get(nil)[1]) do
+                Osi.CharacterRemoveSkill(player, skillName)
+            end
+        elseif character == "Host" then
+            Osi.CharacterRemoveSkill(Osi.CharacterGetHostCharacter(), skillName)
+        else
+            for userProfile, content in pairs(players) do
+                if players[userProfile]["userName"] == character then
+                    Osi.CharacterRemoveSkill(Osi.GetCurrentCharacter(players[userProfile]["userID"]), skillName)
+                end
+            end
+        end
+    else
+        S7_ConfigLog("Error Encountered", "[Error]")
+    end
+end
+
+--  ========================================================
 Ext.RegisterConsoleCommand("S7_AddSkill", S7_AddSkill)
---  ==================================================
+Ext.RegisterConsoleCommand("S7_RemoveSkill", S7_RemoveSkill)
+--  ========================================================
+
+--  #####################################################################################################################################################